Remove radio buttons to toggle display of name
